% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_cor.R
\name{add_cor_stars}
\alias{add_cor_stars}
\title{Add significance stars to a correlations table}
\usage{
add_cor_stars(cor, stars)
}
\arguments{
\item{cor}{A dataframe wit the correlations, as resulting from \code{\link[corrr]{correlate}}}

\item{stars}{A dataframe or matrix with the same dimensions as \code{cor} (omitting the rowname column) of significance stars}
}
\description{
Add significance stars to a correlations table
}

