% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_r.R
\name{ave_r}
\alias{ave_Rsq}
\alias{ave_r}
\title{Calculate average of correlations}
\usage{
ave_r(x, na.rm = TRUE)

ave_Rsq(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Computes the average of the estimates in vector x, using Fisher's r-to-z transformation before computing the mean and reversing the transformation afterward.
\code{ave_Rsq} takes the squareroot of all of the elements first, then applies the transformation, 
computes the mean, reverses the transformation, and squares the result to return the units to a squared correlation.
}
\section{Functions}{
\itemize{
\item \code{ave_Rsq}: Calculate average of squared correlations
}}
\seealso{
\code{\link[base]{atanh}} for computing 
\href{https://en.wikipedia.org/wiki/Fisher_transformation}{Fisher's r-to-z transformation}, 
and \code{\link[base]{mean}}
}

