% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{cat_descriptives_table}
\alias{cat_descriptives_table}
\title{Pretty table for descriptives of categorical variables}
\usage{
cat_descriptives_table(cat_vars, var.names = NULL, caption = NULL,
  show.missing = TRUE)
}
\arguments{
\item{cat_vars}{a dataframe of the categorical variables (factors) to include in the table}

\item{var.names}{an (optional) vector of strings the variable names (will use the column names in cat_vars if none are provided here). 
If var.names is simply "labels" then it will attempt to use variable labels that may have been read in when the dataframe was created (see \code{\link{use_var_labels}}).}

\item{caption}{an (optional) caption to add to the table}
}
\description{
Uses \code{\link[htmlTable]{htmlTable}} to produce nicely formated tables summarizing a set of categorical variables.
}
\examples{
library(dplyr)
warpbreaks \%>\%
select(wool, tension) \%>\%
cat_descriptives_table(var.names=c("Type of wool", "Level of tension"))

}

