% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohen_f.R
\name{cohen.f}
\alias{cohen.f}
\title{Calculation of Cohen's f for ANOVA effect size}
\usage{
cohen.f(means = NULL, n = NULL, sd = NULL)
}
\arguments{
\item{means}{a vector of group means}

\item{n}{sample size in each group}

\item{sd}{standard deviation in each group (also assumed to be the population standard deviation)}
}
\value{
f
}
\description{
Calculates Cohen's f from group means, sample size, and pooled standard deviation for a one-way ANOVA.
Note that this only works for balanced designs (same n in each group).
}
\examples{
cohen.f(means = c(4,6,8), n = 10, sd = 4)
cohen.f(means = c(4,6,8,7,5), n = 10, sd = 4)

library(pwr)
f <- cohen.f(means = c(4,6,8), n = 10, sd = 4)
pwr.anova.test(k = 3, n = 10, f = f)

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). New York:Academic Press.
}
\seealso{
\code{\link[pwr]{pwr.anova.test}}, \code{\link[effsize]{cohen.d}}
}
