% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{cont_descriptives_table}
\alias{cont_descriptives_table}
\title{Continuous descriptives table}
\usage{
cont_descriptives_table(cont_vars, var.names = NULL, caption = NULL,
  show.missing = TRUE)
}
\description{
Continuous descriptives table
}

