% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_section.R
\name{extract_sections}
\alias{extract_sections}
\title{Extract numbered sections from a .Rmd document.}
\usage{
extract_sections(file, sections, dir = getwd(), solutions = TRUE)
}
\arguments{
\item{file}{The .rmd file to extract sections from.}

\item{sections}{The numbers of the sections to extract, as a vector.}

\item{dir}{The directory where \code{file} is saved, and where the extracted file will be written. Defaults to \code{getwd()}.}

\item{solutions}{Logical indicating whether or not to include solutions text. Default is TRUE.}
}
\value{
Does not return anything, but writes a new .rmd file in \code{dir} with the extracted section(s), 
with either "solutions" or "questions" included in the file name and the relevant question numbers. 
Note that if a file of the same name already exists in \code{dir}, it will overwrite it without warning.
}
\description{
This is designed for use with the homework .rmd files for the grad stats sequence in PSY at UO. 
All of the homework problems (question text, data, answer code, and solution text) can be stored in one .rmd file.
This function automatically extracts the desired questions (sections) and prints them in a new .rmd file, optionally including the solutions, preserving the header information from the original file.
This makes it easy for instructors to keep all of the homework material in just one file.
If you wish to update the approach for a particular kind of analysis, you can easily update it in all of the relevant questions at once,
and then quickly re-generate individual .rmd files to send to students.
}
\details{
Note that the formatting of the master .rmd file is very important. 
This function will not work as expected if the .rmd file is not formatted according to these specifications:
Each question must be its own level-1 section, beginning with \code{# } and then the question number, and a period (e.g. \code{# 1.} for the first question).
Note that there must be exactly one space between the \code{#} and the question number, and the question number must be followed by a period. 
Whitespace after the period is optional. 
Sub-parts to the questions must use level-2 or greater headers, but the numbering/lettering system is unimportant (e.g. \code{## a} or \code{## 1.} are both fine for a sub-section header).
All question text must be in bold (e.g., \code{**This is a question.**}). 
You can use bold in answer text, but it cannot be a whole line; for example, \code{This answer requires **emphasis**.} will be considered answer text, 
since the line does not begin and end with \code{**}. 
It is okay if there is whitespace (e.g. tab indenting) before or after the \code{**} that begin and end question lines, but no other characters.

When \code{solutions = TRUE}, the document header wil be included (everything up until the beginning of section 1) 
as will all text from the selected section(s). Any white space, etc. in the file will be preserved.
When \code{solutions = FALSE}, the header is included, but only the section text that is marked as being part of the question (section headers, bolded lines, and certain code chunks).
White space is discarded, and a newline is entered between each markdown line to preserve formatting.
}
\examples{
# extract the question text for questions 1, 2, 5, and 8 from homework_all_solutions.Rmd
extract_sections(file = homework_all_solutions.Rmd, section=c(1,2,5,8), solutions = FALSE)

# extract the corresponding answer key for the above questions
extract_sections(file = homework_all_solutions.Rmd, section=c(1,2,5,8), solutions = TRUE)

}

