% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{is.binary}
\alias{is.binary}
\title{Binary vectors}
\usage{
is.binary(x)
}
\arguments{
\item{x}{the vector to test}
}
\description{
Is a general test of whether a vector is coded like a binary variable, with only 0's and 1's.
}
\examples{
is.binary(mtcars$vs) # TRUE
is.binary(mtcars$gear) # FALSE

# missing values don't prevent it from identifying a vector as binary
is.binary(c(mtcars$vs, NA)) # TRUE

}

