% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_lm.R
\name{print_lm}
\alias{print_lm}
\title{Print regression summary inline}
\usage{
print_lm(model, digits = 2, digits_p = 3)
}
\arguments{
\item{model}{a model object resulting from lm}

\item{digits}{digits to round R² and F statistic to}

\item{digits_p}{digits to round the p value to}
}
\value{
A string with the model results in APA format, like "R2 = .07, F(1, 18) = 1.42, p = .249".
}
\description{
Prints a string that summarizes lm model result in APA format
}
\examples{
print_lm(lm(mpg ~ wt, data = mtcars))

}
