% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_r.R
\name{print_r}
\alias{print_r}
\title{Print r or R-squared value}
\usage{
print_r(r, digits = 2)
}
\arguments{
\item{r}{a number, like from a correlation or model R-squared}

\item{digits}{digits to round to}
}
\description{
Takes a number and returns a string with no leading zero and rounded to digits decimal places.
}
