% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_effects.R
\name{simple_effects}
\alias{simple_effects}
\title{Simple effects tests for factorial ANOVAs}
\usage{
simple_effects(outcome, test, by, data)
}
\arguments{
\item{outcome}{the outcome variable for the models (dependent variable)}

\item{test}{the factor to get simple main effects for}

\item{by}{the factor to split the models by}

\item{data}{a data frame}
}
\value{
An object with classes 'anova' and 'data.frame' that includes tests for the main effect of \code{test} within each level of \code{by}, using the Mean Sq residual from the full model.
}
\description{
Slices an interaction by levels of one factor, \code{by}, testing for the simple main effect of the other factor \code{test} at each level of \code{by}.
This is different from simply splitting the data by \code{by} and then running one-way ANOVAs on each split because \code{simple_effects} uses the Mean Sq residual from the full model.
}
\examples{
# simple main effect of tension at each type of wool
simple_effects(data = warpbreaks, outcome = "breaks", by = "wool", test = "tension")
# simple main effect of wool at each level of tension
simple_effects(data = warpbreaks, outcome = "breaks", by = "tension", test = "wool")

}

