% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_dir.R
\name{source_dir}
\alias{source_dir}
\title{Executes all of the scripts in a directory}
\usage{
source_dir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{directory to source}

\item{trace}{print names of files as they're sourced? default=TRUE}

\item{...}{other arguments to pass to \code{\link[base]{source}}}
}
\value{
None
}
\description{
Uses \code{\link[base]{source}} to run all of the R scripts in a directory. 
A handy way to load a library of functions for a project, for example.
}
\examples{
source_dir("lib")

}
\seealso{
\code{\link[base]{source}}
}

