% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cc.R
\name{test_cc}
\alias{test_cc}
\title{Tests of canonical dimensions}
\usage{
test_cc(cc)
}
\arguments{
\item{cc}{the canonical correlations to test, output from from \code{\link[CCA]{cc}}}
}
\value{
A data frame where each row is one dimension from the cannonical correlation (in order from largest to smallest), with columns for Wilk's Lambda, F statistic, df for the F test, and p value for the F test.
}
\description{
Based closely on the code from \url{http://www.ats.ucla.edu/stat/r/dae/canonical.htm}, with slight modifications to make it work more easily as a function.
}
\examples{
cc1 <- CCA::cc(iris[ , 1:2], iris[ , 3:4])
test_cc(cc1)

}
\references{
\url{http://www.ats.ucla.edu/stat/r/dae/canonical.htm}
}
\seealso{
\code{\link[CCA]{cc}}
}

