% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{use_var_labels}
\alias{use_var_labels}
\title{Use SPSS variable labels as columnn names}
\usage{
use_var_labels(df)
}
\arguments{
\item{df}{A dataframe from either \code{\link[forign]{read.spss}} or \code{\link[haven]{read_spss}}}
}
\value{
A copy of the same dataframe, but with column names replaced with variable labels whereever possible.
}
\description{
Checks for the presence of variable labels from SPSS data files
read with either \code{\link[forign]{read.spss}} or \code{\link[haven]{read_spss}}.
If available, it will overwrite the column names of df with variable names.
For any columns where variable labels are not available, the original column names will be retained.
}

