% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{xtabs_table}
\alias{xtabs_table}
\title{Creates a clean crosstabs table}
\usage{
xtabs_table(xtab, var.names = NULL, caption = NULL)
}
\arguments{
\item{xtab}{a contingency table with class "xtabs"}

\item{var.names}{an optional vector of variable names to use in table}

\item{caption}{an optional caption to include for table}
}
\value{
An \code{\link[htmlTable]{htmlTable}} table, which will appear in the Viewer. 
Display it in your browswer, and then copy-paste it into Word or Libre Office.
}
\description{
Uses \code{\link[htmlTable]{htmlTable}} to make a clean, pretty html-formatted table from a \code{xtabs} object.
}
\examples{
x <- xtabs(~ cyl + gear, data = mtcars)
xtabs_table(x, var.names = c("Number of cylinders", "Number of gears"))

}

