% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{bin_descriptives_table}
\alias{bin_descriptives_table}
\title{Binary descriptives table}
\usage{
bin_descriptives_table(bin_vars, var.names = NULL,
  header = "Percent above threshold", caption = NULL, show.missing = TRUE,
  show.n = FALSE, show.n.sucess = FALSE)
}
\description{
Binary descriptives table
}

