% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{cont_descriptives_table}
\alias{cont_descriptives_table}
\title{Continuous descriptives table}
\usage{
cont_descriptives_table(
  vars,
  var.names = NULL,
  caption = NULL,
  show.missing = TRUE,
  type = "htmlTable",
  ...
)
}
\arguments{
\item{vars}{a dataframe of the variables to include in the table}

\item{var.names}{an (optional) vector of strings the variable names (will use the column names in vars if none are provided here). 
If var.names is simply "labels" then it will attempt to use variable labels that may have been read in when the dataframe was created (see \code{\link{use_var_labels}}).}

\item{caption}{an (optional) caption to add to the table}

\item{show.missing}{logical indicating whether or not to include information on the number of missing cases}

\item{type}{a charcter string specifying which function to use to generate the table 
(\code{\link[htmlTable]{htmlTable}} or \code{\link[pander]{pander}}). Default is "htmlTable".}

\item{...}{additional arguments passed to \code{\link[htmlTable]{htmlTable}} or \code{\link[pander]{pander}}}
}
\description{
Uses \code{\link[htmlTable]{htmlTable}} or \code{\link[pander]{pander}} to produce nicely formated tables summarizing a set of continuous variables.
}
