% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{cont_descriptives_table}
\alias{cont_descriptives_table}
\title{Continuous descriptives table}
\usage{
cont_descriptives_table(cont_vars, var.names = NULL, caption = NULL,
  show.missing = TRUE)
}
\arguments{
\item{var.names}{an (optional) vector of strings the variable names (will use the column names in cat_vars if none are provided here). 
If var.names is simply "labels" then it will attempt to use variable labels that may have been read in when the dataframe was created (see \code{\link{use_var_labels}}).}

\item{caption}{an (optional) caption to add to the table}
}
\description{
Continuous descriptives table
}

