% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{corr_table}
\alias{corr_table}
\title{Correlations table, with optional descriptives}
\usage{
corr_table(cont_vars, var.names = NULL, caption = NULL, plot = FALSE,
  show.means = FALSE, digits = 2, stars = FALSE)
}
\description{
Correlations table, with optional descriptives
}

